Fix questionnaire type mismatches
We had separate types defined in TypeScript and Runtypes. This led to a
TypeScript error when trying to pass a `QuestionnaireCreateInput` into
`QuestionnaireBuilder.graphqlInput`.
