+++ .changeset/clever-cougars-burn.md
@@ -1,0 +1,5 @@
+---
+'wingman-fe': patch
+---
+
+**QuestionnaireBuilder:** Fix TypeScript mismatches
